More checks for time limits and code cleanup.
Also reordering of prune_bounds/theorems to enable earlier time pruning.

+++ src/switches.c
@@ -823,7 +823,7 @@ switcher (const int process, int index, int commandline)
 	}
       else
 	{
-	  set_time_limit (integer_argument ());
+	  set_time_limit ((double) integer_argument ());
 	  return index;
 	}
     }
